requirements.txt
python-telegram-bot==13.15
gTTS
requests
beautifulsoup4
openai
python-dotenv
pytz
win10toast
